アニメーション変更
プレイヤーのをmixamoのに差し替えた

+++ player.h
@@ -21,7 +21,7 @@ private:
 	bool			m_ready;		//構えているか
 	char*			m_OldAnimationChara;	//アニメーションの名前の旧名
 	char*			m_NowAnimationChara;	//現在のアニメーションの名前
-	int m_Frame = 0;				//アニメーションのフレーム
+	float m_Frame = 0;				//アニメーションのフレーム
 	float rate = 0;					//アニメーションブレンドのレート
 
 	//ここに	シェーダー関連の変数を追加
